Test grouping in main.ts
